requirements.txt
fastapi~=0.115.12
uvicorn[standard]>=0.23.2,<0.28.0
sqlalchemy>=2.0.22,<2.1.0
pydantic[email]~=2.11.4
pydantic-settings~=2.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart # Form-data (login uchun)
jinja2>=3.1.2,<3.2.0
python-dotenv>=1.0.0
# psycopg2-binary # Agar PostgreSQL ishlatilsa, kommentni oching va o'rnating

# Celery va Redis uchun
celery~=5.5.2
redis~=6.1.0
flower>=2.0.1,<2.1.0 # Celery monitoring uchun (ixtiyoriy)

# Boshqa kerakli kutubxonalar (agar bo'lsa)
# httpx # Agar API testlari yoki tashqi API chaqiruvlari uchun kerak bo'lsa
# alembic # Ma'lumotlar bazasi migratsiyalari uchun (bu loyihada qo'shilmagan)
eventlet~=0.40.0